define-alias -- first version of new lexically scoped aliases
The only problem is that it can only be used with rash syntax, and
outside of rash syntax gives a syntax error...  I want to be able to
define an alias (eg. cd) that I can access as bare `cd` in rash syntax,
but also be able to access in rash syntax as $cd and more importantly be
able to access OUTSIDE of rash syntax.