save repl readline history to preferences file
like the readline library does, except using a different name.

Actually, what the readline library does is a bit of nonsense.  When
loading it, it saves the new history when exiting, but then saves the
old history on top of it!  I'm now saving the history to the same
preference file, but under a different preference name.

Ideally, I should use a different file.  But this was easy to copy
from what readline normally does, and using the readline library at
all *should* be only a temporary stopgap until I write a better line
editor.